Move spd core implementation into humility-spd
