+++ Cargo.toml
@@ -19,8 +19,9 @@ members = [
     "humility-log",
     "humility-net-core",
     "humility-pmbus",
-    "humility-stack",
     "humility-probes-core",
+    "humility-spd",
+    "humility-stack",
     "cmd/auxflash",
     "cmd/console-proxy",
     "cmd/counters",
@@ -132,6 +133,7 @@ humility-log = { path = "./humility-log" }
 humility-net-core = { path = "./humility-net-core" }
 humility-pmbus = { path = "./humility-pmbus" }
 humility-probes-core = { path = "./humility-probes-core" }
+humility-spd = { path = "./humility-spd" }
 humility-stack = { path = "./humility-stack" }
 cmd-auxflash = { path = "./cmd/auxflash", package = "humility-cmd-auxflash" }
 cmd-console-proxy = { path = "./cmd/console-proxy", package = "humility-cmd-console-proxy" }
