+++ cmd/rendmp/Cargo.toml
@@ -24,6 +24,6 @@ humility-hiffy.workspace = true
 humility-i2c.workspace = true
 humility-idol.workspace = true
 humility-pmbus.workspace = true
+humility-spd.workspace = true
 humility.workspace = true
 
-cmd-spd.workspace = true
